# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "PsicoGestion API",
    "description": "Documentación completa de endpoints con autenticación JWT."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT en formato: Bearer <token>"
    }
  },
...
  "paths": {
...
    "/api/users/becarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar becarios",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Becarios", "schema": { "type": "array", "items": { "$ref": "#/definitions/User" } } } }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar usuarios",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Usuarios", "schema": { "type": "array", "items": { "$ref": "#/definitions/User" } } } }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear usuario",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "body", "name": "body", "schema": { "$ref": "#/definitions/UserCreate" } }],
        "responses": { "201": { "description": "Usuario creado" } }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": ["Usuarios"],
        "summary": "Actualizar usuario",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/UserUpdate" } }],
        "responses": { "200": { "description": "Usuario actualizado" } }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario (soft-delete)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Usuario inactivado" } }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": { "type": "integer", "example": 1 },
        "email": { "type": "string", "example": "test@example.com" },
        "nombre": { "type": "string", "example": "Ana" },
        "apellido": { "type": "string", "example": "López" },
        "telefono": { "type": "string", "example": "5512345678" },
        "rol": { "type": "string", "example": "psicologo" },
        "especialidad": { "type": "string", "example": "Clínica" },
        "activo": { "type": "boolean", "example": true }
      }
    },
    "UserCreate": {
      "type": "object",
      "required": ["nombre", "apellido", "email", "password"],
      "properties": {
        "nombre": { "type": "string" },
        "apellido": { "type": "string" },
        "email": { "type": "string" },
        "telefono": { "type": "string" },
        "rol": { "type": "string" },
        "especialidad": { "type": "string" },
        "fundacion_id": { "type": "integer" },
        "activo": { "type": "boolean" },
        "password": { "type": "string" }
      }
    },
    "UserUpdate": {
      "type": "object",
      "properties": {
        "nombre": { "type": "string" },
        "apellido": { "type": "string" },
        "email": { "type": "string" },
        "telefono": { "type": "string" },
        "rol": { "type": "string" },
        "especialidad": { "type": "string" },
        "fundacion_id": { "type": "integer" },
        "activo": { "type": "boolean" },
        "password": { "type": "string" }
      }
    },
...
  }
}